pyproject.toml
[project]
name = "sts-suite"
version = "0.1.0"
description = "Terminal debugger for Feetech STS3215 smart serial bus servos"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "questionary>=2.0.1",
    "pyserial>=3.5",
    "rustypot>=1.4.2",
    "textual>=0.50.0",
    "textual-plotext>=0.2.1",
]

[project.scripts]
sts = "sts_suite.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sts_suite"]
